+++ CloudShell.RabbitMQ.Client/CloudShellRabbitMQCredentialException.cs
@@ -1,0 +1,14 @@
+namespace CloudShell.RabbitMQ.Client;
+
+public sealed class CloudShellRabbitMQCredentialException : Exception
+{
+    public CloudShellRabbitMQCredentialException(string message)
+        : base(message)
+    {
+    }
+
+    public CloudShellRabbitMQCredentialException(string message, Exception innerException)
+        : base(message, innerException)
+    {
+    }
+}